formatChicagoShortDateKey: render full month name, not abbreviated
The "Short" in the name means "no weekday", not "abbreviated month" —
naming convention differs from formatChicagoShortDate (the instant
variant), which also shortens the month to three letters. Keep full
month names for client-facing email copy; preserves the pre-fix
rendering on email.ts:777 ("April 20, 2026" not "Apr 20, 2026").

Docstring updated to call out the naming asymmetry so future readers
don't mirror the wrong variant.

+++ src/lib/dateChicago.ts
@@ -162,9 +162,14 @@ export function formatChicagoLongDateKey(d: Date): string {
 
 /**
  * Compact human-readable date for a UTC-midnight date-key — e.g.
- * `"May 20, 2026"`. Same input contract as `formatChicagoLongDateKey`
- * (date-keys only, never real instants); drop the weekday for inline
- * list views or short sentences embedded in paragraphs.
+ * `"April 20, 2026"` or `"May 20, 2026"`. Same input contract as
+ * `formatChicagoLongDateKey` (date-keys only, never real instants);
+ * drop the weekday for inline list views or short sentences embedded
+ * in paragraphs.
+ *
+ * Short form = no weekday. Month name is still spelled out in full,
+ * unlike `formatChicagoShortDate` (the instant variant) which also
+ * abbreviates the month to three letters.
  *
  * Same `timeZone: "UTC"` rationale as `formatChicagoLongDateKey` — see
  * that helper and the file header for the instant-vs-date-key split.
@@ -172,7 +177,7 @@ export function formatChicagoLongDateKey(d: Date): string {
  */
 export function formatChicagoShortDateKey(d: Date): string {
   return d.toLocaleDateString("en-US", {
-    month: "short",
+    month: "long",
     day: "numeric",
     year: "numeric",
     timeZone: "UTC",
